Decompose CLI argument parsing module
