+++ crates/cli/src/frontend/arguments/bandwidth.rs
@@ -1,0 +1,7 @@
+use std::ffi::OsString;
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub(crate) enum BandwidthArgument {
+    Limit(OsString),
+    Disabled,
+}
