+++ crates/cli/src/frontend/arguments/env.rs
@@ -1,0 +1,23 @@
+use std::env;
+
+pub(crate) fn env_protect_args_default() -> Option<bool> {
+    let value = env::var_os("RSYNC_PROTECT_ARGS")?;
+    if value.is_empty() {
+        return Some(true);
+    }
+
+    let normalized = value.to_string_lossy();
+    let trimmed = normalized.trim();
+
+    if trimmed.is_empty() {
+        Some(true)
+    } else if trimmed.eq_ignore_ascii_case("0")
+        || trimmed.eq_ignore_ascii_case("no")
+        || trimmed.eq_ignore_ascii_case("false")
+        || trimmed.eq_ignore_ascii_case("off")
+    {
+        Some(false)
+    } else {
+        Some(true)
+    }
+}
