+++ crates/cli/src/frontend/arguments/mod.rs
@@ -1,0 +1,11 @@
+mod bandwidth;
+mod env;
+mod parsed_args;
+mod parser;
+mod program_name;
+
+pub(crate) use bandwidth::BandwidthArgument;
+pub(crate) use env::env_protect_args_default;
+pub(crate) use parsed_args::ParsedArgs;
+pub(crate) use parser::parse_args;
+pub(crate) use program_name::{ProgramName, detect_program_name};
